backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
asyncpg==0.30.0
redis==5.2.1
python-jose[cryptography]==3.4.0
pydantic-settings==2.8.1
